requirements.txt
openai==1.55.3 
httpx==0.27.2
streamlit==1.34.0
pyyaml==6.0.1
neo4j==5.17.0
geopandas==0.13.2
shapely==2.0.3
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
cachetools==5.3.3
duckdb==0.10.2
libpysal==4.10
esda==2.5.1
python-dotenv==1.0.1
pysal==2.6
spaghetti==1.7.4
contextily==1.4.0
pointpats==2.4.0
mapclassify>=2.6
fiona==1.8.22
pyproj==3.4.1